fix: ignore pytest progress markers in failed summary parsing
- only treat short-summary FAILED lines as actionable test failures
- add regression coverage for progress output like FAILED [ 31%]
- keeps untested-code.py findings focused on real pytest failures
